+++ core/src/main/java/me/lozm/domain/auth/vo/AuthorizationVo.java
@@ -3,7 +3,7 @@ package me.lozm.domain.auth.vo;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import me.lozm.global.common.BaseVo;
+import me.lozm.global.object.vo.BaseVo;
 
 @Getter
 @SuperBuilder
